refactor(pricing-detailed-analysis): extract sub-components into separate files
Split pricing-detailed-analysis into PricingOverviewCards,
KitBreakdownTable, PeriodComparisonTable, and InsightsPanel to improve
readability and maintainability.
